adding a new sample function with diagnostics
