+++ src/Samplers/BouncyParticleSamplers.jl
@@ -86,8 +86,9 @@ end  # mutable struct BPS
 function BPSAD(dim::Int, U::Function; refresh_rate::Float64=0.0, grid_size::Int=10, tmax::Union{Float64, Int}=2.0, 
                     vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, AD_backend::String="ForwardDiff")
 
-  ∇U = set_AD_backend(AD_backend, U, dim)
+  ad_setup = set_AD_backend(AD_backend, U, dim)
+  ∇U = ad_setup.gradient
 
   return BPS(dim, ∇U, refresh_rate=refresh_rate, grid_size=grid_size, tmax=tmax, 
-                    signed_bound=signed_bound, adaptive=adaptive, AD_backend=AD_backend)
+                    signed_bound=signed_bound, adaptive=adaptive, AD_backend=ad_setup.AD_backend)
 end
