+++ src/Samplers/ZigZagSamplers.jl
@@ -116,11 +116,12 @@ mutable struct ZigZag{G,KF,KR,KRV,KSR,KSRV,KVJ} <: AbstractPDMP
 end
 
 function ZigZagAD(dim::Int, U::Function; refresh_rate::Float64=0.0, grid_size::Int=10, tmax::Union{Float64, Int}=2.0, 
-                    vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, AD_backend::String="Zygote")
+                    vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, AD_backend::String="ForwardDiff")
 
-    ∇U = set_AD_backend(AD_backend, U, dim)
+    ad_setup = set_AD_backend(AD_backend, U, dim)
+    ∇U = ad_setup.gradient
 
     return ZigZag(dim, ∇U, refresh_rate=refresh_rate, grid_size=grid_size, tmax=tmax, vectorized_bound=vectorized_bound, 
-                    signed_bound=signed_bound, adaptive=adaptive, AD_backend=AD_backend)
+                    signed_bound=signed_bound, adaptive=adaptive, AD_backend=ad_setup.AD_backend)
 end
